Fix bug when initial weights are integer valued
This bug fix casts any user defined initial weights,
passed as a Numpy array of integers,
as floats.  Otherwise, using the += operation with an
integer Numpy array and a float Numpy array will raise
an error.

+++ logistic_regression.py
@@ -104,7 +104,7 @@ def batch_gradient_ascent(train_inputs, train_targets, initial_weights=None,
     if initial_weights is None:
         weights = np.random.randn(train_inputs.shape[1] + 1)
     else:
-        weights = initial_weights
+        weights = initial_weights.astype(np.float64)
 
     train_costs = []
     test_costs = []
